feat: externalize hardcoded strings and refactor UpdateAvailableDialog
