+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,15 @@
 <resources>
     <string name="app_name">Gemini Pro</string>
+    <string name="new_update_available">New Update Available</string>
+    <string name="update_dialog_message">Version %1$s is available.\nWould you like to download it?</string>
+    <string name="whats_new">What\'s New:</string>
+    <string name="skip_this_version">Skip this version</string>
+    <string name="remind_me_later">Remind me later</string>
+    <string name="download">Download</string>
+    <string name="show_diagram">Show Diagram</string>
+    <string name="preview_html">Preview HTML</string>
+    <string name="tap_and_hold_video">Tap and hold the video to download it</string>
+    <string name="no_video_found">No video found at this location</string>
+    <string name="link_copied">Link copied!</string>
+    <string name="no_url_to_copy">No URL to copy.</string>
 </resources>
